backend/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
mysqlclient==1.4.2
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
